Workaround for crash at the end of test
Due to incorrect sequence of HPURegister and RecipeSingletion
destructors, we see crash at the end of every test/use case.
Crash occurs when HPURegistrar destructor is invoked and it goes
on to access the already destroyed RecipeSingleton object.

This patch is a ugly workaround, needed to unblock the
pytorch CI.

The proper fix for this issue is tracked in .

Change-Id: I8b80c499cc70f3fe16dc2d36eeb05952b8f037f0
Signed-off-by: Internal Developer <internal-developer@intel.com>
Signed-off-by: Internal Developer <internal-developer@intel.com>