Graph mode: Propagate layout information/Handle Channel Last
In fused op graph, if the output of an op is an intermediate tensor,
then permute is not added to the output. However, if the consumer op for
this output is layout agnostic, then no permute gets added to the graph.
Support for channel last layout is also added.

For example, in the graph below, the %output.1 from convolution needs a permute,
but its not added as its an internal tensor. The output from relu, %input.4
needs to add the permute.

Pass the input layout info to output edges only for ops that have single input
and single output, and the output is layout agnostic.

  graph(%0 : Float(*, *, *, *),
    %1 : Float(*, *, *, *),
    %2 : Float):
    %3 : int = prim::Constant[value=1]()
    %4 : bool = prim::Constant[value=0]()
    %5 : int[] = prim::Constant[value=[0, 0]]()
    %6 : int[] = prim::Constant[value=[1, 1]]()
    %output.1 : Float(*, *, *, *)
      = aten::convolution_overrideable(%0, %1, %2, %6, %5, %6, %4, %5, %3) 
    %input.4 : Float(*, *, *, *) = aten::relu(%output.1) 
    return (%input.4)

Change-Id: I92083400c892035ecd63883829487ccebc80f8c7
Signed-off-by: Internal Developer <internal-developer@intel.com>
Signed-off-by: Internal Developer <internal-developer@intel.com>