Tools/Resnet: Set Epoch number in Tensor Probe
So far tensor probe dumped data in directories corresponding to
iteration numbers. But if tensor probe is used across epochs,
the iteration numbers will not be unique. So we need to add another
level of directory for epochs. This is also useful when loading from a
checkpoint.

Also added documentation on tensor probe env variables

Change-Id: I3757b477d82f1864550d8a4bc203486f2c9ab2c6
Signed-off-by: Internal Developer <internal-developer@intel.com>