Remove tensor ptr from kernels and use reference.
Tensor lifetime management of the tensor that are part
of the execution recipe is done by taking reference
of the Tensor and pass it to event so that tensors are
prevented from deallocating till the execution is completed.
For enabling event for stream synchronization need
to get the tensor references.

Change-Id: I36b89f2b6ead38ce8736d2b8f077b250f3b7d757
Signed-off-by: Internal Developer <internal-developer@intel.com>