Fix crash when enabling caching in grapmode
Enabling graph mode caching results in syncLaunch
failure with address overlap.
The issue was because the intermediate aten tensor
were getting deleted when first iteration completes,
which resulted in using address pointer which was freed
causing the syncLaunch failure.

Fix by storing the reference of intermediate tensor
in recipe cache.

Change-Id: I32ddea02ce4f57b78a92df956bf75be3f6664ebb
Signed-off-by: Internal Developer <internal-developer@intel.com>
Signed-off-by: Internal Developer <internal-developer@intel.com>