[WORKAROUND] Resnet: Drop last batch if it is of partial batch size
The last batch will be a partial batch if the data set size is not a multiple
of batch size. This can cause re-compilation of the graph resulting
in memory alloc failure due to increased memory usage.
Work around this issue with the Dataloader option to drop the last partial
batch until a formal fix is available.

Change-Id: If46f05a6788bd7709d39bbe438f63b2e9f11931d
Signed-off-by: Internal Developer <internal-developer@intel.com>