Separated traced modules for train and eval
The traced module for train and eval steps need to be different.
This is because:
Batchnorm has different flow in train and eval phase.
If in graphmode, we trace the flow only in train phase,the eval phase
flow wont be captured.And when the graph executes in eval phase, it
will still go through the train phase flow. This results in improper
loss values in evaluation pahse.

Change-Id: I9c99672e0d2cf5e0d5e933b6139818af1e85448f
Signed-off-by: Internal Developer <internal-developer@intel.com>