Support for output tensor handle release for fused graph
Create the output tensor on a recipe cache hit. This allows
the cached recipe to be usable for any static graph matching
the saved graph, without tying it to a specific subgraph
instance.

The option is disabled by default with JIT graph mode. For
lazy mode, it is enabled by default.

To enable creation of output tensor on a cache hit, use
HABANA_PGM_ENABLE_CACHE=15

To disable creation of output tensor on a cache hit, use
HABANA_PGM_ENABLE_CACHE=1

Change-Id: I4c0cb4843ea1cbc8eba4906c1c51957ad3eabd59
Signed-off-by: Internal Developer <internal-developer@intel.com>