Minor cleanup for blocking copy and recipe
moved the handling of event for the recipe to a
common function instead of repicating it in compile
and execute recipe.
Instead of passing the recipe struct to the event
callback in the recipe launch event, pass only
the recipe ptr which should not be deleted while
execution is inflight and pass the pt_input in the
callback as value, so that it automaticlly creates
the tensor refrence in PT.
For blocking call, removed the tensor ref from
the event callback.

Change-Id: I80d7bb6a9c5da9b7ce5dc7031ca1a8d4087df2f1
Signed-off-by: Internal Developer <internal-developer@intel.com>