Refactoring of addcdiv in Adam Optimizer for BERT
This patch refactors the addcdiv used in Adam optimizer in BERT for
getting better cache hits. Additionally this refactoring avoids certain
failures related to timing issues observed in BERT.

The cache hit (and hence avoidance of recompilation) is due to the use
of step size (=learning rate) as a tensor in the recatored code. Without
this kind of refctoring, the stepsize was passed to pytorch habana kernel
as a scalefactor 'value'. since the learning rate changes for every
iteration in BERT with SQUAD/MRPC, this resulted in cache miss. When step
size is passed as a tensor, though the learning rate still changes, tensor
can be patched into the recipe and the change vale DMA-ed like any other
tensor.

Change-Id: I186d0a5f03feb7ab9fea64f97ce7bc4232caae59
Signed-off-by: Internal Developer <internal-developer@intel.com>