[REVERTME] BERT: Model modification to avoid nn.Dropout() if dropout prob is zero
This is done to avoid dropout for debug/convergence test purposes.
Passing the dropout probability as 0.0 in SQUAD/MRPC training script is
sufficient to avoid call to nn.Dropout in Eagermode. This is not enough
in trace mode. Hence the change at model level.

Modification is done in BERT and MRPC models.
(SQUAD Does not use Dropout in the fine tuning stage)

Revert after implementing direct dropout interface when dropout can be
bypassed in tracemode without model change.

Change-Id: I8a265c039668c6b9ad642d674a70012cf2c7bdf0
Signed-off-by: Internal Developer <internal-developer@intel.com>