Keep input pytorch tensor in lazy IR nodes
PyTorch tensors that are input to the lazy IR graph need to have
a lifetime at least as long as the graph is not executed. Since
the PyTorch tensors are created and sent back to the framework/model,
there is a chance of the tensor getting deleted by the framework/model
before the graph is executed.

To resolve this, store the PyTorch tensors that are input to the lazy
IR graph within the IR node. This allows the PyTorch tensor lifetime
to be managed by the IR graph itself. Once the IR graph is executed,
the IR nodes get deleted and the input PyTorch tensors are also removed.

Instead of storing it in the hpu::input nodes, the PyTorch tensors are
stored in IR nodes that consume the hpu::input node. This is because storing
the PyTorch tesnor in hpu::input node will introduce a circle -

  PT Tensor -> Lazy Tensor -> Data -> ir_value -> hpu::input node --
    ^                                                              |
    |                                                              |
    ----------------------------------------------------------------

Instead, it is stored in the node consuming hpu::input, for example -

  permute(PT Tensor)

This will be stored as -

  PT Tensor -> Lazy Tensor -> Data -> ir_value -> hpu::input node
    ^                                                  ^
    |                                                  |
    |                                            permute IR node
    |                                                  |
    ----------------------------------------------------

Change-Id: Ie704c227e2429dcfd5142977c8238b582e0f4c72
Signed-off-by: Internal Developer <internal-developer@intel.com>