[Sw-0] Fix Infinte looping in case of OOM
Recipe is removed from recipe cache when OOM
is hit, there will be one recipe be a atleast one
used recipe in case when the failure is during the
creation of output tensor in graph mode. Fix it
by returing recipe count removing the used recipes
from the total count.

When ASYNC flag is set to Zero, the recipe active count
usage is not valid as the execution is sequential. Remove
the usage in launch in graph mode.

Change-Id: Ib403f87a6a7d7f7c414762521588b0f642689242
Signed-off-by: Internal Developer <internal-developer@intel.com>