log_livemem_alloc fix for peak memory calculation
The synDeviceMalloc/Free logging from synapse can face a race
condition as stated below -

If allocations and free happen from multiple threads, then
it is possible to have a scenario when the free followed by an
allocation is seen by the logger in reverse order and the
freed ptr is returned by alloc. It needs to be checked if a free
follows this message with the same ptr, which then is most likely
due to the logger seeing the free followed by alloc in revsere
order as alloc followed by free.

In such case, keep track of such duplicates and adjust the memory
peak count accordingly.

Change-Id: I6f7a1871e2f133fa6c4b28e591b3ba4c4b37f031
Signed-off-by: Internal Developer <internal-developer@intel.com>