Move hbLazy to hbtorch module
Since we dont need any python level interfacing code for lazy, we can
reuse hb_torch module instead of introducing a new python module for
lazy.

And move lazy_executor.{h,cpp} appropriately in preparation of removing
hblazy dir. hblazy module is left as is right now as CI builds it. It
will be removed after it is removed from the build script.

Additionally cleanup of unused arguments is included in this change.
Change-Id: I67fa15d5a2d301ce64fc349e0c849e36ebf90688