Fix crash when using dynamic workspace
- when Dynamic workspace is used the allocation of workspace
happen in launch, so the active_recipe_count check need to
include the current recipe which is getting launched count >1.

- when recipe ws is requested with bigger size than the current
then the WS is freed and reallocated. if the recipe execution is
in flight deletion of WS will caused unexpected behaviour. So
check and wait for the recipes to finish execution for freeing the
current WS buffer.

Change-Id: I8dface5291dcf51e75f070b293f02b8c8854ffb3
Signed-off-by: Internal Developer <internal-developer@intel.com>