BERT:PT:Add support for grad accumulation steps > 1
The grad acc steps commandline arg should be converted to a
tensor so that it can be used to average the loss value.

Also restored the grad accumulation steps values in
scripts/run_pretraining.sh to the defaults of 128 and 512 for
phase 1 and 2 respectively. The train batch sizes have also
been increased such that batch size per processing unit is
32 and 8 for phase 1 and 2 respectively.

Change-Id: Idc65d819f8ae86d29f47bfa273d831eb5abb6880
Signed-off-by: Internal Developer <internal-developer@intel.com>