Release device before PyTorch process exit
Change the creation order of synapse OSAL and bridge
HPRegistrar object

Create the synapse_helpers::device before creatng the
HPRegistrar object. For this to work, use a static flag
to check whether the device have been created already or not.

If the device isn't created, then create it first, resulting
in OSAL object creation.

After this, create the HPURegistrar object and insert the
device. This ensures that the destruction order is ~HPRegistrar
followed by ~OSAL, which ensures correct destruction of devices.

Apart from this, to handle the thread_local objects from synapse
and their destruction order, a thread_local object is created
in the bridge that drives the correct order of destruction of
synapse deivces before the thread_local objects from synapse
are destroyed.

Change-Id: Ib32f23fee3b37e7f166a4afaa4172563b582840f
Signed-off-by: Internal Developer <internal-developer@intel.com>