Avoid memcopies in lazy mode BN
BN RMV calculations need to be done inplace.
But since underlying TPC kernel doesnt support inplace update, memcopy nodes are used for torch script and eager modes.
But these can be removed with a trick in Lazy mode as it involves front end and back end tensors.
The RMV calculations can be done out of place and the tensors can be attached to the right front end tensor.
In addition training and evaluation operators are separated to ease
debugging and readability

Change-Id: Ic13873cbb55a271f1fd1d1ec780b1a6333ee7df1
Signed-off-by: Internal Developer <internal-developer@intel.com>