Optimize Coalesced Pooling Search
- Used pool list to have only the free chunks instead
of both used and unused chunks.
- if reuse chunk is not available, then get the next
best available size for splitting instead of skipping
because of size check.
- Fix merge previous function to return the correct chunk,
 Instead of returning the zero sized chunk.
- Remove recursive lock and use the mutex lock as
it not required in the current call scenario and is
expensive.
- Removed cached recipe removal in case of OOM when
request is for WS as removing the cache recipe will only
adds up host memory and the deleteing intermediate tensor
without copying back the value to host will cause incorrect
value.

Change-Id: I4a12bd32450a4af9ae81af77b80a7fcc25346943
Signed-off-by: Internal Developer <internal-developer@intel.com>