Iter0 loss values mismatch issue compared to trace mode
Its observed that in lazy mode with FusedSGD and BN-Relu fusion, the loss values mismatch doesnt happens. The reason could be without this two features, the graph is big and GC will be splitting graphs differently. With this extra mark step, the loss is same for eager, trace and lazy.

Change-Id: I8801d96e1776eafce8927e5c927d542fb6f5cad3
Signed-off-by: Internal Developer <internal-developer@intel.com>