Lazy conserves the PT tensor dtype
Keeping the dtype intact (in case of Long/Double) on the PT tensor
so that the framework sees the original type. The Internal tensor
is with the new type and the ops will work with this internal type.

For a D2D to long/double, we do not haev these types available in
the device. Hence, the copy target is the source target still and
the D2H does the required conversion.

copy_hpu_ has been modified to handle float to double D2H.

Change-Id: I9c1f45ecd1b7cf8e77ee6a9842369a1b5864a426
Signed-off-by: Internal Developer <internal-developer@intel.com>