Fix empty tensor name in patching table
Patching table in the synLaunch has empty tensor
names, the names were getting empty sometimes when
copy contructor of habana tensor() is used.
the name were getting empty string as with libc++
string move constructor does empty the source. it
copies all of the bytes of the source, and then zeros
all of the bytes of the source.

Change-Id: I72de308296807e2f23968dee22f06eda6274b5fe
Signed-off-by: Internal Developer <internal-developer@intel.com>