Identify uninitialized ir::value weak_ptr
Check the owner_before for an empty weak pointer.
As per https://en.cppreference.com/w/cpp/memory/weak_ptr/owner_before,

"The order is such that two smart pointers compare equivalent only if
they are both empty or if they both own the same object"

If the weak_ptr is uninitialized, expired() call still returns true as
the use_count() is 0 and we can't differentiate an uninitialized tensor
against an initialized and expired tensor.

The owner_before with an empty weak_ptr is going to return false if the
m_data_ptr is uninitialized.

Change-Id: I33663b298c19e126cb4ac72eac92b7a5f7dd5358
Signed-off-by: Internal Developer <internal-developer@intel.com>