Fix perf degrade in case of using virtual handle
In Bert, one of the output tensor gets deleted before the
recipe execution is completed. if the next request happens
to copy and if the released memory handle is reused, this
leads to event mapping to the compute and copy stream, which
results in copy stream to wait for the compute stream to
complete which resulted in increased time for copy.
In the recipe launch call back, hold the reference of output
tensors so that tensor is alive till the execution is complete.

Change-Id: I2ae7c156f647e834383b2257bb8d9bfab97d2673
Signed-off-by: Internal Developer <internal-developer@intel.com>