Fix crash if in case of pending events to sync
when there are pending events to sync and if the
process exit, the device_cleanup() is getting called
without flushing the events, this causes crash with
as device is deleted. in this case destructor is
called as~HPURegistrarPerThreadTracker->device.cleanup()
->~device->~stream->~HPUDeviceAllocator, this causes
stream sync thread to fail as device is already deleted.
so call FlushStream() in cleanup to make sure the
stream sync is completed before device, stream deletion.

Change-Id: I928afa216c328a0de348a971c51ad7e3e3bfdf98
Signed-off-by: Internal Developer <internal-developer@intel.com>