fixed NaN issue in masked_fill kernel
the kernel was lowered into 2xmul and an add
which caused NaNs when filling inf value.
Moved to use where_fwd kernel instead which is more
suitable and does not have this side effect.
Since the tpc where does not support in-place, added
a memcpy to copy the output back to the self tensor

Change-Id: Id44e861325869944c7fa12ea0a78272d8a60c28b