Create JIT stack without all stack inputs
While creating the JIT stack on a first level IR cache miss, pass the
entire stack inputs, and not after removing duplicates. This is required
as the po_data.inputs are withthe duplicates and the post order graph also
has the nodes for all duplicates. The permute pass looks at both the inputs
and we need to pass all the stack inputs for it to work.

Change-Id: Id1b8e151c8a82082291a2bdbf0aa6b6c58c3b461
Signed-off-by: Internal Developer <internal-developer@intel.com>