masked_fill bf16 support and input dtype check fix
fixed input dtype check for masked_fill_ to be per tensor - the mask
should be an int/char/byte/bool. the data was narrowed down to the tpc
supported dtypes (float,bf16,int).

moved to use scalar_to_device_tensor to convert the scalar to tensor and
removed the previously added method that did the same.

added unit test for bf16 data.

Change-Id: Ib13c732f71f829c362415a1112adbb93f6e9d568