Fix for invalid ptr when deleting device memory
The size of the memptr was getting zero in the
chunks list becasue chunk size to split and
requested size was same. This was caused to
to inccorect check before splitting the chunks
in one of call sequence.Fix is to check the chunk
size is greater than requested then do split chunks.

Change-Id: I23da0220c28c4cc4faba11ba6924020fa1a82552
Signed-off-by: Internal Developer <internal-developer@intel.com>