Avoid unnecessary applyPendingGraph calls for H2D
When a tensor is sent for DMA from host to device, a typical example
being .to(hpu), currently the GetHbLazyTensorData() ends up calling
applyPendingGraph() unnecessarily. This increases time taken for H2D
DMAs. If a model is sending ops with scalars and they are converted
to tensors during op accumulation, the time adds up to lazy op
accumulation duration.

The applyPendingGraph() should only be required when there is a real
graph with ops to be executed.

Change-Id: Ifc90761552ba0656f311e8d3043315e91e948276
Signed-off-by: Internal Developer <internal-developer@intel.com>