is_alias_of lazy implementation
Override is_alias_of from PT framework and return result
based on view_table information. This patch has the actual
call disabled as it depends on the as_strided framework to be
added.

The functionality depends on view table information to be
maintained within the bridge. Hence, this patch asserts that
PT_HPU_ENABLE_VIEW_TABLE is set to true.

Change-Id: I7699aa9d7ff496fa4cb79ffc4a3f695939654e35
Signed-off-by: Internal Developer <internal-developer@intel.com>