Ad-hoc HCL groups (workaround)
Implementation of a temp workaround for HCL subgroups:
Creating a world on the fly before every collective op
and destroying it after its completion.

Adhoc groups are enabled if env var PT_HPU_ENABLE_ADHOC_GROUPS
is set to true. It is also expected, that the calling
process sets HCL_CONFIG_PATH appropriately defining
the world to be created.

This workaround must be removed once dynamic HCL groups
are implemented.

Change-Id: I003b055ec40df10f54d2d3ee9f3b277d37d4afa9