Remove linking of synapse logger in habana helper
synapse logger is used only for tracing and dumping data,
inculding this by default to habana_pytorch_plugin will
enable logs based on the default flag. So removed the
usage of synapse logger in habana helper.

Added device release to dump data when signal_handler
is called.

Added function to get the synapse library path for
the calculating the hash for cache version

Change-Id: I33f18b770390ca2906ba8f35df4e7f53e0d2b67c
Signed-off-by: Internal Developer <internal-developer@intel.com>