Deadlock avoidance with GIL
Media Pipe acquires GIL and requests for data_ptr. At the same
instant, the autograd is also kicked in which is in need of
GIL. Both the instances makes use of the bridge lock and there
is a cyclic dependancy with GIL leading to deadlock.

This change aims to use the lock only in context creation and
not during data_ptr fetch

Change-Id: Id70b30d212a7d782fee74b94f3f5e13fe1dc6236
Signed-off-by: Internal Developer <internal-developer@intel.com>