Use truncation while casting Float to int
This is what CPU does. Hence needed to match CPU results.
Use round to nearest even in case of fp32 to bf16.

This patch handles the float to int conversioni(which goes via
CastOperator path) issue in JIRA .

TODO: float to long conversion issue in the same JIRA.
This goes via the memcopy path and needs a different fix.

Change-Id: I94dbfb73272a16d3b9227a718562f972e44609be
Signed-off-by: Internal Developer <internal-developer@intel.com>