Collectives on views
If the collective operation happens on a view output, then subsequent access to the tensor doesnt pick the updated value.
This is because view operations are added 'lazily' on torch ops and collective is not part of torch op.

The fix is to treat collectives as inplace operation i.e. add a strided_insert node to update the parent
Also support views on point to point communication.

Change-Id: I2256ba828ffe713cdb6ec948a8cde98aa60c7d2a
Signed-off-by: Internal Developer <internal-developer@intel.com>