Add mark step before lazy H2D copy
This is initial fix for a inplace H2D tensor copy:
In case there is a pending graph on that tensor,
it will be executed _after_ teh copy, this overwriting
the tensor.

This scenario is typical to model load from a checkopoint.

It should be optimized to check if indeed there is
a pending graph on the dest tensor ().

Change-Id: I9e63af1d9339bcafcbfca9e86da6f0b23340640a