Add support to get dataptr for default pool
In case of Stringent Pool allocator, device memory will
be allocated only when recipe or copy operation.
For media pipe, tensor data ptr is used to copy the
data to the device and this needs device memory
to be allocated when htcore get_data_ptr is queried.
So added support to return device memory when using
htcore.get_data_ptr().

Change-Id: I40cbe1170b3d6ea6362719573d0e6f48b146c3dd
Signed-off-by: Internal Developer <internal-developer@intel.com>