cleanup device in ~device_context
In case of distributed and lazy mode, the recipes
can be in queue when model script exit happens,
this will cause hccl delete and this causes the
current stream in FIFO to hang. So before destroying
of device_context, call clean-up of device which
will call flush stream before destroying the
hccl.

Change-Id: Ie8b861dc23019d1deb43fc4a39e5d5717c1b68a3
Signed-off-by: Internal Developer <internal-developer@intel.com>