Remove stack_out op implementation
It is a dispatch=true; default=true op.
So not necessary to be implemented.
And since  is fixed, stack_out op can be removed.

Also removed erasing of pt_inputs vector in CatOut op.
Erasure was causing a seg fault in lazy mode( but not in
eager via lazy) because pt_inputs happened to be empty
in AllocateAndAddAynapseNode of CatOut op.
Since Erasure of pt outputs is not considered needed,
this was removed.

TODO: Review the need/handling of pt_inputs in general.

Change-Id: I5112e5a9f2664c727bda1e48f73110bb9995119f
Signed-off-by: Internal Developer <internal-developer@intel.com>