Pass scalar as double to get_tensor_for_scalar()
get_tensor_for_scalar() took float as the alpha(i.e. scalar)
value. This meant that all the scalar values to be cached needed
to be converted to float irrespective of actual data type of scalar.
But this can cause data integrity issues in the following cases:
1. If the actual scalar is an integer type and its value can not be
   represented accurately by float32 i.e, if the actual scalar could
   have been represented in int32 and its value is more than 2^24 and
   it is not a power of 2.
  (same is applicable for int64 values beyond 2^24 and  not a power of 2.
  But since we can not handle integers beyond int32 in HW as of now,
  the useable range is same as that of int32.)
2. if scalar value is actually a double value that can not be represented
   by float. This must be an unlikely case. In any case we can not reperesent
   floating point numbers beyond float32 in the HW now. So we cant anyway
   handle this. This case can be ignored for now.

This patch introduces a qiuck fix to take scalar as double.
This fixes issue 1 above for int32 scalars and int64 within int32 range.

This fixes issue 2 at the get_tensor_for_scalar() interface, but since
we convert double to float tensor, issue2 cant be solved which is ok
for now.

Change-Id: I9c2bc7b4fd71e2c595b479125108be2412d555b8
Signed-off-by: Internal Developer <internal-developer@intel.com>