Handle send/recv permuted tensors
Sending a tensor doesn't have metadata field,
hence we can't send the info if tensor is dense or permuted.
So for first functional step, we'll always permute it back to be dnese before sending it.
In future it can be optimized if we can send metadata too
via send mechanism to provide this info.

Change-Id: Ie7f2d36826b6848d59a5cf0d3abec40f55b366d1