Handle send/recv permuted tensors for SFG
Sending a tensor doesn't have metadata field,
hence we can't send the info if tensor is dense or permuted.
Marking false synTensorSetAllowPermutation for tensor send as part of lowering to Syanpse.
If during lazy accum send tensor already permuted, we make it dense,
using inplace clone and marking it as synTensorSetAllowPermutation false.

Change-Id: I3097c52b8fa6d65c8e2777cbca39a43c6b6b8068