Enable caching for addcdiv/addcmul variants
addcdiv and addcmul are defined as :
out = input + value * (other1 op other2), where 'op' is div or
mul and 'value' is a scalar.

This patch converts 'value' from scalar to tensor to avoid graph
cache misses when its value changes. Further the resulting tensor
is cached for DMA transfer to avoid redundant DMAs.

Change-Id: Ia2032a2e05d1759a3450a5892c9a5607dd84a6bf
Signed-off-by: Internal Developer <internal-developer@intel.com>