Fix for defragmentor issue
In case when defragmenter is run when workspace
OOM happened, has a issue with invalid memory ptr.
There was duplicate entry for workspace when creating
memory map. So added check to avoid duplicate.

When defragmenter is run, sometimes synLaunch failed
with address overlap, since defragmenter moves the
device address, the locked address needs to be updated
as it can have stale address.

Fix updating of workspace size when defragmenter is run.

Change-Id: Ifae5a124b4436099ad63907bd2af9f3056fd2c10