FusedLamb: Get rid of multiple [1.0f] allocations in py
This patch alters FusedLamb and its corresponding PyTorch operators to
prevent redundand allocations of [1.0f] trust ratio for each of the
model parameter with each optimizer's step. Instead, [1.0f] tensors are
allocated as non-persistent tensors directly in the phase2 operator.

This reduces the host time of MLPerf BERT PT2 (PyTorch/Gaudi2) by 20..40
ms. Even though the effective step time is not affected, we wanted to
reduce the host time to give a safe room to mitigate potential hazards
when running BERT on MLPerf submission machines.

Thanks-To: knagasubramanian@habana.ai
Thanks-To: vgoel@habana.ai
Thanks-To: gsingh@habana.ai
Thanks-To: jayachandranb@habana.ai

Change-Id: I053b2406c0b6bc87eadf9e43b8a3d324a2a4945a