SmoothL1/HuberLoss:pass i8 as condition tensor for where guid
TPC doc mandates that the condition tensor be int8.
But PT kernel was passing a type dependent on the input dtype
(like bf16). This was causing mismatch issues with CPU.

This came as a regression due to certain CGUID related changes.
(ref: )
Earlier, even though PT was passing a dtype other than int8,
GC was adding a cast to convert it to the dtype required by TPC.
With the CGUID change, this cast is no longer applied and hence
it showed up as a regression.

Fixed by passing the condition tensor as int8(by setting int8 as
the output dtype of the opertion that creates the condition tensor)

Change-Id: I77a06c30ee8cd5981961c1d7c4e1b11f25a0c3f5
Signed-off-by: Internal Developer <internal-developer@intel.com>