Seed setting for Randperm operator
Randperm operator uses randomshuffle guid which takes in
a seed to randomize the permutation. However the seed setting
was not working correctly resulting in lack of randomization
in permute(Rand perm generated the same permuted seq. always)

Fixed the seed setting. Generated the seed tensor at FE and
passed on to BE with an associated schema change.
Also did some cleanup to remove unused legacy eager randperm
implementation.

Change-Id: I4651bf7480250425bb8c8c319bbb2026c6b69d4a
Signed-off-by: Internal Developer <internal-developer@intel.com>