Fix workspace shrink issue
-Use current workspace size, instead of ws_usage_map last
element. Using ws_usage_map, does not point to current workspace
used. map is ordered and last elements point to the largest size
so far used.
-Correct the grapInfo log with the required WS and the actual
used.
-Add additional check in get_least_worksapce() to pick the least
workspace Based on the memory availability. sometimes the workspace
difference is small with respect to the one which is currently used
and can lead to OOM.

Change-Id: I9649e536defc3c081e8ace291b1a1e0330e77e74
Signed-off-by: Internal Developer <internal-developer@intel.com>