index_select() fixes: Handle repeating indices and large Index tensor
The following cases which were not previously handled are now handled:
1. If there are repeating(duplicate) elements in index tensor,
  the corresponding source entries are to be added as many times
  to self at the index i as the index i repeats in the index tensor.
2. Since index tensor can have repeating elements, the index tensor
  size at the selected dim can be larger than the self size at that dim.
  This was causing a crash previously. With the curent change this case
  is also handled.

Also,
TPC mandates indices (and hence the source as well) to be sorted for
scatter_add_fwd guid to work accurately. So added sorting.

TODO: If Gaudi2 can support scatter_add without sorting, add a check and
avoid sorting on G2

Note: Inorder not to risk any immediate regression, the new implementation
is invoked only in the case where scenario 2 above is satisfied.
This impl. will be made default for all cases after sufficient testing
is done.

Change-Id: Ic6533feec81028e7d030d0b47b2ade3fd5fc8913
Signed-off-by: Internal Developer <internal-developer@intel.com>