FusedAdamW: Fix device attribute of state variable tensors
Fused AdamW kept a state variable for device set based on the
device attribute applicable at the time of calling init().
Later, the device attribute of some of the  state variable tensors
and other tensors were set based on this stored device value.
This assumed that by the time the optimizer init() was called, the model
had been moved to the target device. However, this need not be true
always. Model may be on CPU when optimizer init is called and when
the model is later moved to HPU, the state variable for device will
still remain CPU and hence the device attribute of some other
tensors will get set to CPU instead of to HPU.

Remove the state variable for device and set device attribute
based on the device applicable at the time of creation of the
respective tensors.

This patch also avoids calling optimizer operations if there are no
grad tensors(i.e if the weight tensor list is empty)

Change-Id: Ia7171a6063eaadfbed922b5c4b379ecb8dd00927
Signed-off-by: Internal Developer <internal-developer@intel.com>