Fix issue with memory lock in collective
In collective, lock_address are called separetly
for input and output address and this leads locking
of memory address and causes defrgmenter to fail
in case where the output address allocation
requires defragmention of memory for which the
collective is not executed.
So need to call lock address for both input/output
address together to execute the collective.

Change-Id: I028b19a8c0719862536981871023bc0a4592529c
Signed-off-by: Internal Developer <internal-developer@intel.com>
Signed-off-by: Internal Developer <internal-developer@intel.com>