compile and launch thread for lazy eager
In lazy eager we need to separate the execute/launch
in a separate thread.
Reaching per op lazy eager latency targets would require
overlapping the lowering in the bridge with the compilation
in Synapse for the previous op. This will require pipelining
the graph compilation and launch in a separate thread.

Change-Id: I9bddb08126512908b4b285ce9527f5e623b9e817
Signed-off-by: Internal Developer <internal-developer@intel.com>
Signed-off-by: Internal Developer <internal-developer@intel.com>