Modify CPU fallback check logic for index kernels
The existing check was causing execution to be directed to HPU
(i.e, execution did not fall back to CPU) in certain cases where
index tensor did not have value. This resulted in runtime error
"bad optional access" when accessing those tensors HPU bridge/Op.

Modified the check to run the op on HPU only if
all the optional tensors in the index tensor list are
have value and are defined.

Change-Id: Iaee82600a8fbab07e0a726ad021ca035d57f4401
Signed-off-by: Internal Developer <internal-developer@intel.com>