Non-lazy backend tensor for pipelined lowering
Provides a backend tensor for the lowering flow in non-lazy eager.
The non-lazy eager kernels require to send the input and output tensors to the pipelined
lowering flow in PT Habana bridge.
The input and output tensors need to be disassociated with the frontend tensors available
to PT framework, to avoid race conditions on tensor metadata content. It also cleanly
separates the backend container that is sent to synapse lowering.

Change-Id: I2409b33272821385376f8d6360df657b061e1e3c
Signed-off-by: Internal Developer <internal-developer@intel.com>